requirements.txt
mysql-connector-python==9.5.0
Pillow==12.1.0
python-dotenv==1.2.1
python-telegram-bot==22.5
